Moved tests to xUnit on .NET Core.
