+++ ConsoleFramework.sln
@@ -11,7 +11,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleFramework", "Console
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ManyControls", "ExamplesStandalone\ManyControls\ManyControls.csproj", "{20E4164D-D86A-4042-B1BE-353E21CFDA06}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestProject1", "TestProject1\TestProject1.csproj", "{1579CF00-4DFB-48E0-B6C0-098DA0FC9F41}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{C19D960E-D8DC-442E-9BC4-8FB1085F72F5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -44,12 +44,12 @@ Global
 		{20E4164D-D86A-4042-B1BE-353E21CFDA06}.DebugMono|Any CPU.Build.0 = Debug|Any CPU
 		{20E4164D-D86A-4042-B1BE-353E21CFDA06}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{20E4164D-D86A-4042-B1BE-353E21CFDA06}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1579CF00-4DFB-48E0-B6C0-098DA0FC9F41}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1579CF00-4DFB-48E0-B6C0-098DA0FC9F41}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1579CF00-4DFB-48E0-B6C0-098DA0FC9F41}.DebugMono|Any CPU.ActiveCfg = DebugMono|Any CPU
-		{1579CF00-4DFB-48E0-B6C0-098DA0FC9F41}.DebugMono|Any CPU.Build.0 = DebugMono|Any CPU
-		{1579CF00-4DFB-48E0-B6C0-098DA0FC9F41}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1579CF00-4DFB-48E0-B6C0-098DA0FC9F41}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C19D960E-D8DC-442E-9BC4-8FB1085F72F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C19D960E-D8DC-442E-9BC4-8FB1085F72F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C19D960E-D8DC-442E-9BC4-8FB1085F72F5}.DebugMono|Any CPU.ActiveCfg = Debug|Any CPU
+		{C19D960E-D8DC-442E-9BC4-8FB1085F72F5}.DebugMono|Any CPU.Build.0 = Debug|Any CPU
+		{C19D960E-D8DC-442E-9BC4-8FB1085F72F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C19D960E-D8DC-442E-9BC4-8FB1085F72F5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
