--- TestProject1/Xaml/EnumsTest/ObjectToCreate.xml
@@ -1,3 +1,0 @@
-﻿<ObjectToCreate>
-  <ObjectToCreate.MyEnum>Variant2</ObjectToCreate.MyEnum>
-</ObjectToCreate>
