--- TestProject1/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="System.Reflection.TypeExtensions" version="4.4.0" targetFramework="net461" />
-</packages>
