+++ Tests/Xaml/EnumsTest/ObjectToCreate.xml
@@ -1,0 +1,3 @@
+﻿<ObjectToCreate>
+  <ObjectToCreate.MyEnum>Variant2</ObjectToCreate.MyEnum>
+</ObjectToCreate>
